+++ src/constants.ts
@@ -7,7 +7,6 @@ export enum Algorithms {
   PBKDF2 = 'PBKDF2'
 }
 export const AES_KEY_LENGTH = 256
-export const PBKDF2_ITERATIONS = 110000
 export const RSA_PUBLIC_EXPONENT = new Uint8Array([1, 0, 1])
 
 export const enum Formats {