+++ src/rsa.ts
@@ -1,4 +1,3 @@
-import { deriveKey } from './aes'
 import { Algorithms, Hashes, Formats, RSA_PUBLIC_EXPONENT } from './constants'
 import { crypto } from './globals'
 import { encode, ABconcat, decode } from './base64'
@@ -26,21 +25,18 @@ export function generateKeypair(keySize: number): PromiseLike<CryptoKeyPair> {
  */
 export async function wrapPrivateKey(
   privateKey: CryptoKey,
-  passphrase: string,
-  PBKDF2salt: Uint8Array,
-  algorithm: 'AES-GCM'|'AES-CBC'
+  wrappingKey: CryptoKey
 ): Promise<string> {
   // Derive a secure key of the specified algorithm using the passphrase and salt
-  const tempKey = await deriveKey(algorithm, passphrase, PBKDF2salt)
   const iv = makeSalt(12)
 
   // Encrypt the private key using the derived key
   const encrypted = await crypto.subtle.wrapKey(
     Formats.PKCS8,
     privateKey,
-    tempKey,
+    wrappingKey,
     {
-      name: algorithm,
+      name: Algorithms.AES_GCM,
       iv
     }
   )
@@ -48,37 +44,29 @@ export async function wrapPrivateKey(
   const concatenated = ABconcat(iv, encrypted)
 
   // Encode to base64 and prepend the encryption algorithm
-  return `${algorithm}:${encode(concatenated)}`
+  return encode(concatenated)
 }
 
 /**
- * Decrypt an RSA private key using the same passphrase and salt that were passed to wrapPrivateKey
+ * Decrypt an RSA key using unwrappingKey (AES-GCM)
  */
 export async function unwrapPrivateKey(
   encodedPrivateKey: string,
-  passphrase: string,
-  PBKDF2salt: Uint8Array
+  unwrappingKey: CryptoKey
 ): Promise<CryptoKey> {
   // Parse the algorithm from the key
-  const [algorithm, concatenated] = encodedPrivateKey.split(':')
-  if (!(algorithm === Algorithms.AES_GCM || algorithm === Algorithms.AES_CBC)) {
-    throw new TypeError('Invalid or malformed private key encoding (unable to parse a valid decryption algorithm).')
-  }
-
-  // Recreate the same key derivation as what occured during the key's wrapping
-  const tempKey = await deriveKey(algorithm, passphrase, PBKDF2salt)
 
   // Decode the provided information and split into iv and encrypted private key
-  const decoded = decode(concatenated)
+  const decoded = decode(encodedPrivateKey)
   const iv = decoded.slice(0, 12)
   const encrypted = decoded.slice(12)
 
   return crypto.subtle.unwrapKey(
     Formats.PKCS8,
     encrypted,
-    tempKey,
+    unwrappingKey,
     {
-      name: algorithm,
+      name: Algorithms.AES_GCM,
       iv
     },
     {
@@ -133,18 +121,21 @@ export async function wrapKey(
     }
   )
 
-  return `${key.algorithm.name}:${encode(new Uint8Array(buf))}`
+  return encode(new Uint8Array(buf))
 }
 
+/**
+ * Unwrap (decrypt) a CryptoKey using an RSA private key
+ */
 export function unwrapKey(
   encodedKey: string,
-  unwrappingKey: CryptoKey
+  unwrappingKey: CryptoKey,
+  // Algorithm of the key to be unwrapped
+  algorithm: 'AES-GCM'|'AEC-CBC' = 'AES-GCM'
 ): PromiseLike<CryptoKey> {
-  const [algorithm, encoded] = encodedKey.split(':')
-
   return crypto.subtle.unwrapKey(
     Formats.Raw,
-    decode(encoded),
+    decode(encodedKey),
     unwrappingKey,
     {
       name: Algorithms.RSA // Algorithm used to unwrap the key